refactor: Organize project structure by features
This commit reorganizes the project directory structure to align with the features described in the README: Data Collection, Trading/Models, and Execution.

- Moved data provider logic to `src/data_collection`.
- Moved backtester and strategy logic to `src/trading`.
- Created placeholder directories for future AI models (`src/trading/models`) and trade execution (`src/execution`).
- Renamed modules for clarity and consistency.
- Updated all import paths to reflect the new structure.
- Cleaned up empty and unnecessary (__pycache__) directories.

The functionality of the backtester remains unchanged, as confirmed by running the example script. This new structure improves modularity and makes the codebase more scalable and maintainable.
