+++ src/trading/backtester.py
