+++ src/trading/strategies/simple_ma.py
